PLANET-5125 Add script to update config
* Based on https://github.com/greenpeace/planet4-circleci-config/blob/6869c0b1754a13521676fa3b9f248472486fa5b3/create_configs.sh.
So the some of this commit message mainly applies to the changes made
to that script.
* Use dockerize and same template as NRO generator to create the new
config file.
* Don't push to Github unless specified in the arguments. Otherwise
still create the tmp config file but don't remove it.
* Put site update logic in a function.
* Add list for develop only sites.
* Add arguments whether to include dev and/or prod.
* Add script to check if config.yml contains any differences compared to
artifacts.yml, which is the file containing the environment variables.
We should probably get rid of the duplication of these variables, as now
you're able to change config.yml without applying the same changes to
artifacts.yml. If we wouldn't check for this when we push new
configuration to all repos, these changes would be overwritten.
* List sites in separate files: dev_sites.txt, test_prod_sites.txt and
prod_sites.txt