PLANET-5409 Build 1 image for stag and prod
* This image was already effictively the same in both workflows. They
used a different set of env vars, but the one it actually needed was
the same for stag and prod, the Google project id. So I created a new
set which only has that env var and the build just works.
* I kept the google project id in the stag and prod env var sets. This
avoids having to change a whole lot of config manually in all NRO repos.
* Merging stag and prod workflows turned out to be surprisingly simple.
In both, the deploy step was using an image tagged with the circleci
workflow run. So just moving the production jobs into the staging
workflow was enough to make it use the same image.
* Rename staging to release where appropriate.
* More specific names for workflows definitions.
* Make hold-production depend on deploy-staging, which makes the
pipeline fit into one screen.